Open Graph: make sure transients are used to save image IDs.
In #6297, we added transients to cache some of the fallback images in jetpack_og_get_image().
However, transients can only be 45 chars maximum.
To work around that limitation, we now create a 44-char string using md5, based on the image URL.

Also added back changes to Site Icon brought in #6303, and accidentally reverted in #6297.